+++ src/main/java/com/savaz/delivery/service/Service.java
@@ -4,6 +4,6 @@ import com.savaz.delivery.model.entity.bean.OrderBean;
 
 public interface Service {
 
-    void topUpAccount (int userId, double amount);
+    void topUpAccount (int userId, long amount);
     void payCheck (int userId, OrderBean orderBean);
 }
